% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc200.R
\docType{data}
\name{cc200}
\alias{cc200}
\title{Automated probabilistic reconstruction of white-matter pathways}
\format{
A data.frame with 11341 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (sagittal)}
  \item{label}{label name from Freesurfer}
  \item{atlas}{name of the atlas}
}
}
\usage{
cc200
}
\description{
Automated probabilistic reconstruction of white-matter pathways
}
\examples{
data(cc200)
}
\references{
Yendiki et al. (2011) Automated probabilistic reconstruction of 
white-matter pathways in health and disease using an atlas of the underlying anatomy.  
Front. Neuroinform. 5:23. doi: 10.3389/fninf.2011.00023
(\href{https://www.ncbi.nlm.nih.gov/pubmed/22016733}{PubMed})
}
\concept{ggseg_atlases}
\keyword{datasets}
